requirements.txt
numpy
pysdl2
pysdl2-dll
PyOpenGL